refactor(cross-chain): move HttpClient types to core/interfaces
