+++ packages/cross-chain/src/core/utils/httpClient.ts
@@ -6,33 +6,14 @@
  * @see https://www.npmjs.com/package/ses
  */
 
+import type {
+    HttpClientConfig,
+    HttpRequestOptions,
+    HttpResponse,
+} from "../interfaces/httpClient.interface.js";
 import { HttpError } from "../errors/HttpError.exception.js";
 import { HttpNetworkError } from "../errors/HttpNetworkError.exception.js";
 import { HttpTimeout } from "../errors/HttpTimeout.exception.js";
-
-export interface HttpResponse<T = unknown> {
-    status: number;
-    data: T;
-    headers: Headers;
-}
-
-export interface HttpRequestOptions {
-    method?: "GET" | "POST" | "PUT" | "PATCH" | "DELETE";
-    headers?: Record<string, string>;
-    /** Objects are JSON-serialized and `Content-Type: application/json` is set. Strings are sent as-is. */
-    body?: unknown;
-    /** Appended to the URL via `URLSearchParams`. Values are stringified; `null`/`undefined` are skipped. */
-    params?: Record<string, unknown>;
-    /** Aborts and throws `HttpTimeout` after this many ms. */
-    timeout?: number;
-    signal?: AbortSignal;
-}
-
-export interface HttpClientConfig {
-    baseURL?: string;
-    headers?: Record<string, string>;
-    timeout?: number;
-}
 
 /** HTTP client with a minimal axios-like surface. */
 export class HttpClient {
